Move density matrices.
