+++ gqcp/include/Operator/SecondQuantized/SQOneElectronOperator.hpp
@@ -20,12 +20,12 @@
 
 #include "Basis/Transformations/JacobiRotationParameters.hpp"
 #include "Basis/Transformations/TransformationMatrix.hpp"
+#include "DensityMatrix//OneDM.hpp"
+#include "DensityMatrix//TwoDM.hpp"
 #include "Mathematical/Functions/CartesianGTO.hpp"
 #include "Mathematical/Functions/LinearCombination.hpp"
 #include "Mathematical/Functions/ScalarFunction.hpp"
 #include "Mathematical/Representation/QCMatrix.hpp"
-#include "Processing/DensityMatrices/OneDM.hpp"
-#include "Processing/DensityMatrices/TwoDM.hpp"
 #include "Utilities/type_traits.hpp"
 
 #include <array>
