+++ gqcp/include/QCMethod/HF/GHF/GHFSCFEnvironment.hpp
@@ -19,9 +19,9 @@
 
 
 #include "Basis/Transformations/TransformationMatrix.hpp"
+#include "DensityMatrix//OneDM.hpp"
 #include "Mathematical/Representation/QCMatrix.hpp"
 #include "Operator/SecondQuantized/SQHamiltonian.hpp"
-#include "Processing/DensityMatrices/OneDM.hpp"
 
 #include <Eigen/Dense>
