+++ gqcp/include/gqcp.hpp
@@ -59,6 +59,10 @@
 #include "Basis/Transformations/SpinResolvedTransformationMatrix.hpp"
 #include "Basis/Transformations/TransformationMatrix.hpp"
 #include "Basis/Transformations/transform.hpp"
+#include "DensityMatrix//OneDM.hpp"
+#include "DensityMatrix//SpinResolvedOneDM.hpp"
+#include "DensityMatrix//SpinResolvedTwoDM.hpp"
+#include "DensityMatrix//TwoDM.hpp"
 #include "Mathematical/Algorithm/Algorithm.hpp"
 #include "Mathematical/Algorithm/CompoundConvergenceCriterion.hpp"
 #include "Mathematical/Algorithm/ConvergenceCriterion.hpp"
@@ -155,10 +159,6 @@
 #include "Operator/SecondQuantized/USQHamiltonian.hpp"
 #include "Operator/SecondQuantized/USQOneElectronOperator.hpp"
 #include "Operator/SecondQuantized/USQTwoElectronOperator.hpp"
-#include "Processing/DensityMatrices/OneDM.hpp"
-#include "Processing/DensityMatrices/SpinResolvedOneDM.hpp"
-#include "Processing/DensityMatrices/SpinResolvedTwoDM.hpp"
-#include "Processing/DensityMatrices/TwoDM.hpp"
 #include "Processing/Properties/BaseElectricalResponseSolver.hpp"
 #include "Processing/Properties/CIElectricalResponseSolver.hpp"
 #include "Processing/Properties/RHFElectricalResponseSolver.hpp"
