+++ gqcp/src/Processing/CIDMCalculators/BaseSpinResolvedFrozenDMCalculator.cpp
@@ -15,7 +15,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with GQCG-GQCP.  If not, see <http://www.gnu.org/licenses/>.
 
-#include "Processing/DensityMatrices/CIDMCalculators/BaseSpinResolvedFrozenDMCalculator.hpp"
+#include "DensityMatrix//CIDMCalculators/BaseSpinResolvedFrozenDMCalculator.hpp"
 
 #include "Utilities/linalg.hpp"
