+++ gqcp/tests/CMakeLists.txt
@@ -1,6 +1,7 @@
 set(test_target_sources)
 
 add_subdirectory(Basis)
+add_subdirectory(DensityMatrix)
 add_subdirectory(Mathematical)
 add_subdirectory(Molecule)
 add_subdirectory(ONVBasis)
