+++ gqcp/tests/Processing/CMakeLists.txt
@@ -1,4 +1,3 @@
-add_subdirectory(DensityMatrices)
 add_subdirectory(Properties)
 
 set(test_target_sources ${test_target_sources} PARENT_SCOPE)
