+++ gqcpy/src/CMakeLists.txt
@@ -1,4 +1,5 @@
 add_subdirectory(Basis)
+add_subdirectory(DensityMatrix)
 add_subdirectory(Mathematical)
 add_subdirectory(Molecule)
 add_subdirectory(ONVBasis)
