+++ gqcpy/src/DensityMatrix/OneDM_bindings.cpp
@@ -15,7 +15,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with GQCG-GQCP.  If not, see <http://www.gnu.org/licenses/>.
 
-#include "Processing/DensityMatrices/OneDM.hpp"
+#include "DensityMatrix//OneDM.hpp"
 
 #include <pybind11/eigen.h>
 #include <pybind11/pybind11.h>
