+++ gqcpy/src/DensityMatrix/SpinResolvedTwoDM_bindings.cpp
@@ -15,7 +15,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with GQCG-GQCP.  If not, see <http://www.gnu.org/licenses/>.
 
-#include "Processing/DensityMatrices/SpinResolvedTwoDM.hpp"
+#include "DensityMatrix//SpinResolvedTwoDM.hpp"
 
 #include <pybind11/eigen.h>
 #include <pybind11/pybind11.h>
