+++ gqcpy/src/Processing/CMakeLists.txt
@@ -1,4 +1,3 @@
-add_subdirectory(DensityMatrices)
 add_subdirectory(Properties)
 
 set(python_bindings_sources ${python_bindings_sources} PARENT_SCOPE)
